src/requirements.txt
python-dotenv
requests-pkcs12
streamlit
streamlit-keycloak
SQLAlchemy
psycopg2-binary